src/requirements.txt
Flask-PyMongo>=2.3.0
PyMongo>=4.3.3
Flask>=2.2.3
fhir.resources>=6.5.0
requests==2.26.0
flask-unittest>=0.1.3
